+++ NEWS.md
@@ -1,5 +1,7 @@
 # ellmer (development version)
 
+* `chat_bedrock()` gains `api_args` argument (@billsanto, #295).
+
 * New `content_pdf_file()` and `content_pdf_url()` allow you to upload PDFs to supported models. Models that currently support PDFs are Google Gemini and Claude Anthropic. With help from @walkerke and @andrie (#265).
 
 * `Chat$get_model()` returns the model name (#299).